refactor: improve update_user internal logic
